requirements.txt
Pillow
PyYAML
setuptools
setuptools
tensorboardX
torch
torchvision
FrEIA